add sample.size to conf

+++ ql/src/java/org/apache/hadoop/hive/ql/abm/udaf/SrvAvgComputation.java
@@ -14,7 +14,7 @@ public class SrvAvgComputation extends UDAFComputation {
   protected double baseSsum = 0;
   protected int baseCnt = 0;
 
-  protected int N = 0;
+  protected long N = 0;
   protected int groupCnt = -1;
 
   protected double currentSum = 0;
@@ -24,7 +24,7 @@ public class SrvAvgComputation extends UDAFComputation {
   protected double confidenceLower = Double.POSITIVE_INFINITY;
   protected double confidenceUpper = Double.NEGATIVE_INFINITY;
 
-  public void setTotalNumber(int N) {
+  public void setTotalNumber(long N) {
     this.N = N;
   }
 
@@ -42,7 +42,7 @@ public class SrvAvgComputation extends UDAFComputation {
     this.currentCnt = 0;
     this.groupCnt ++;
   }
-  
+
   public void clear() {
     result.clear();
     doubleMatrix.clear();
